modified the main.rb file and board class
